memento/backend/requirements.txt
# Flask backend dependencies
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
bcrypt==4.1.2
pyjwt==2.8.0
python-dotenv==1.0.0

# Face recognition dependencies
insightface==0.7.3
opencv-python==4.8.1.78
numpy==1.24.3
onnxruntime==1.16.3

# Additional dependencies for insightface
protobuf<5
albumentations>=1.0.3
scikit-image>=0.19.3
Cython>=0.29.32
easydict>=1.10

# Optional but recommended for performance
Pillow>=9.0.0